Be more explicit about boolean value parsing
Supported boolean values are now t, T, true, TRUE, f, F, false, and
FALSE.  This is what the strconv.ParseBool function supports with
the exception of 1 and 0.  1 and 0 would be parsed as ints in the
line protocol.

Previously, any non-true value would be parsed as false.  e.g.
value=blah would parse to false.  This will now return an error as
parsing time.