Fix data race in tsm dataFile
dataFile was not protected by a mutex which causes a data race and live
code and tests.  filesAndLock used reflect.DeepEqual on a copy of dataFile
slices.  reflect.DeepEqual appears to access unexported dataFile fields
which can't be protected.  This was changed to use a equals func that will
require a mutex to be acquired.

The other issue was that many of the dataFile funcs access the mmap without
acquiring a lock.  When a dataFile is deleted (possibly during rewriting),
reads from the mmap could return invalid data because references to the dataFile
are still in use by other goroutines.

Fixes #4534